fix(api): opaque fire engine delete + poll interval (#2443)

+++ apps/api/src/scraper/scrapeURL/engines/fire-engine/delete.ts
@@ -1,5 +1,4 @@
 import { Logger } from "winston";
-import * as Sentry from "@sentry/node";
 
 import { robustFetch } from "../../lib/fetch";
 import { MockState } from "../../lib/mock";
@@ -7,21 +6,25 @@ import { fireEngineStagingURL, fireEngineURL } from "./scrape";
 
 export async function fireEngineDelete(
   logger: Logger,
-  jobId: string,
+  jobId: string | undefined,
   mock: MockState | null,
   abort?: AbortSignal,
   production = true,
 ) {
+  // jobId only supplied if we need to defer deletion
+  if (!jobId) {
+    logger.debug("Fire Engine job id not supplied, skipping delete");
+    return;
+  }
+
   await robustFetch({
     url: `${production ? fireEngineURL : fireEngineStagingURL}/scrape/${jobId}`,
     method: "DELETE",
     headers: {},
-    ignoreResponse: true,
-    ignoreFailure: true,
     logger: logger.child({ method: "fireEngineDelete/robustFetch", jobId }),
     mock,
     abort,
   });
 
-  // We do not care whether this fails or not.
+  logger.debug("Deleted job from Fire Engine", { jobId });
 }
